+++ src/touchElementP.ts
@@ -1,3 +1,4 @@
+import { ParseSelector } from 'typed-query-selector/parser'
 import { touchElement } from './touchElement'
 
 /**
@@ -11,9 +12,32 @@ import { touchElement } from './touchElement'
  *
  * @example
  * ```typescript
- * const button = await touchElementP<HTMLButtonElement>('#my-button')
+ * // -------------------------------------------------------------------------
+ * // Automatically attempts to parse CSS selectors into element types, which
+ * // should work for tag-qualified CSS selectors
+ * // -------------------------------------------------------------------------
+ *
+ * touchElementP('form.my-form button#my-button')
+ * // Promise<HTMLButtonElement>
+ *
+ * // -------------------------------------------------------------------------
+ * // When using non-recognised selectors the element type defaults to `Element`
+ * // -------------------------------------------------------------------------
+ *
+ * touchElementP<HTMLButtonElement>('#my-button')
+ * // Promise<Element>
+ *
+ * // -------------------------------------------------------------------------
+ * // When it fails to infer the element types from given CSS selector you can
+ * // specify the type explicitly
+ * // -------------------------------------------------------------------------
+ *
+ * touchElementP<HTMLButtonElement>('#my-button')
+ * // Promise<HTMLButtonElement>
  * ```
  */
-export function touchElementP<T extends Element>(selector: string, scope: ParentNode = document): Promise<T> {
-    return new Promise((resolve) => touchElement<T>(selector, resolve, scope))
+export function touchElementP<S extends string>(selector: S, scope?: ParentNode): Promise<ParseSelector<S>>
+export function touchElementP<T extends Element>(selector: string, scope?: ParentNode): Promise<T>
+export function touchElementP(selector: string, scope: ParentNode = document) {
+    return new Promise((resolve) => touchElement(selector, resolve, scope))
 }